+++ src/datamodule/aenet/data_set.py
@@ -58,8 +58,8 @@ class StructureDataset(object):
 		scale = [ ]
 		for iesp in range(self.N_species):
 
-			sh = torch.Tensor(sfval_avg[iesp]).double()
-			sc = 1/torch.sqrt(torch.Tensor(sfval_cov[iesp]).double() - sh**2)
+			sh = torch.Tensor(sfval_avg[iesp])  
+			sc = 1/torch.sqrt(torch.Tensor(sfval_cov[iesp])   - sh**2)
 
 			# Check if scale is finite. If infite it means that some of the values are always 0
 			sc[sc == float("Inf")] = 0.0
